refactor: Split out property settings functionality for sharing
