+++ Source/Contrib/SimulatorTester/UserSettings.cs
@@ -16,7 +16,7 @@
 // along with Open Rails.  If not, see <http://www.gnu.org/licenses/>.
 
 using System.Collections.Generic;
-using ORTS.Settings;
+using ORTS.Common;
 
 namespace SimulatorTester
 {
